creation de la balle avec la partie graphique de la balle

+++ src/brique.cpp
@@ -1,6 +1,6 @@
 #include "brique.h"
 
-brique::brique(int x, int y, double largeur, double hauteur) : formeRectangle{x, y, largeur, hauteur}
+brique::brique(int x, int y, int largeur, int hauteur) : formeRectangle{x, y, largeur, hauteur}
 {
 
 }
